refactor(middleware): correct middlewares connected with socket
